refactor: moved methods to new class
